Docker: update build.sh (#22)
* Added an additional expect condition for pacstrapping

An additional -exact entry is required for a prompt that states "Do you want to skip the above package for this upgrade? [y/N]" to supply the answer with a "Y". The "... above package ..." here refers to the ones in $PKGIGNORE.

* Added sed, pacman and procps-ng

sed, pacman and procps-ng need to be installed during pacstrap as they are required by build-helper.sh script but do not get installed otherwise. Therefore, adding them to $PACMAN_EXTRA_PKGS.

* Update build.sh